+++ frontend/web/src/i18n/messages/en-US/index.js
@@ -3,6 +3,7 @@ import auth from "./auth.js";
 import common from "./common.js";
 import devices from "./devices.js";
 import groupControl from "./group-control.js";
+import logs from "./logs.js";
 import settings from "./settings.js";
 
 export default {
@@ -11,5 +12,6 @@ export default {
   ...auth,
   ...devices,
   ...groupControl,
+  ...logs,
   ...app,
 };